LayoutTests/platform/mac/media/encrypted-media/fps-generateRequest-expected.txt

RUN(promise = navigator.requestMediaKeySystemAccess("com.apple.fps", capabilities))
Promise resolved OK
RUN(promise = access.createMediaKeys())
Promise resolved OK
RUN(keys.setServerCertificate(serverCertificate))
RUN(session = keys.createSession())
EXPECTED (session != 'null') OK
RUN(promise = session.generateRequest("sinf", initData))
Promise resolved OK
END OF TEST
